Create README.md
Abstract—In recent years web applications that are hacked every day estimated to be 30 000, and in most cases, web developers or website owners do not even have enough knowledge about what is happening on their sites. Web hackers can use many attacks to gain entry or compromise legitimate web applications, they can also fool people with the utilization of phishing sites to gather their delicate data such as usernames, passwords, and credit card numbers. In response to this, the need is raised to take proper measures to understand the risks and be aware of the vulnerabilities that may affect the website and hence the normal business flow. In this study, mitigations are set against the most common web application attacks, and the web administrator is enabled to detect phishing links, which is a social engineering attack. The detection is performed by various machine learning algorithms and deep learning techniques. The study also shows web application logs generation that simplify the process of analyzing abnormal user’s behavior to indicate when behavior is out of bounds, out of scope, or against the rules. The developed application has been tested and evaluated against various attacks scenarios, the results obtained from the test process showed that the website had successfully mitigated these attacks. In order to find the best technique to detect phishing links, a comparison is made between various algorithms, and the one that produced the highest test accuracy is chosen.